requirements.txt
asgiref==3.7.2
Babel==2.13.1
Brotli==1.1.0
certifi==2023.7.22
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==41.0.5
cssselect2==0.7.0
Django==5.0.1
django-admin-interface==0.28.0
django-auth-ldap==4.6.0
django-colorfield==0.11.0
django-debug-toolbar==4.2.0
django-environ==0.11.2
django-parler==2.3
django-renderpdf==4.0.0
django-rosetta==0.10.0
django-simple-history==3.4.0
django-webpush==0.3.5
docutils==0.20.1
environ==1.0
fontawesomefree==6.5.1
fonttools==4.44.0
html5lib==1.1
http-ece==1.1.0
idna==3.4
joblib==1.3.2
nltk==3.8.1
numpy==1.26.2
pandas==2.1.4
parler==1.0.1
Pillow==10.1.0
polib==1.2.0
psycopg==3.1.12
psycopg2-binary==2.9.9
py-vapid==1.9.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydyf==0.8.0
pyphen==0.14.0
python-dateutil==2.8.2
python-ldap==3.4.3
python-slugify==8.0.1
pytz==2023.3
pywebpush==1.9.4
regex==2023.12.25
reportlab==4.0.7
requests==2.31.0
scipy==1.11.4
six==1.16.0
sqlparse==0.4.4
stringcase==1.2.0
text-unidecode==1.3
tinycss2==1.2.1
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
weasyprint==60.1
webencodings==0.5.1
zopfli==0.2.3
